[MINOR][TESTS] Move TestEnvHelper to core test sources, scope to SparkTestSuite
Re-targets the trait at `SparkTestSuite` only, dropping the shared usage
across `LogKeysSuite` (common-utils) and `ConnectFunSuite`
(connect-client-jvm). The trait now lives in `core/src/test/scala/`
beside `SparkTestSuite`, so `TestEnvHelper.class` ships in the `core`
test-jar that every downstream test module already depends on — no
Maven dependency-resolution surprises.

Reverts:
- `LogKeysSuite` back to its local `getWorkspaceFilePath` and
  `regenerateGoldenFiles` (common-utils does not depend on core).
- `ConnectFunSuite` back to its local `getWorkspaceFilePath`
  (connect-client-jvm does not depend on core, and closure
  serialization to the connect server made cross-classpath sharing
  awkward).
- `PlanGenerationTestSuite`'s `private val regenerateGoldenFiles`.
- `RemoteSparkSession.syncTestDependencies` no longer uploads
  common-utils test-classes; `IntegrationTestUtils.commonUtilsTestClassDir`
  removed.

Net effect: `SparkTestSuite` defers two helpers to the trait; everyone
else is unchanged from master.

Generated-by: Claude Code (Opus 4.7)
