+++ sql/connect/client/jvm/src/test/scala/org/apache/spark/sql/connect/test/ConnectFunSuite.scala
@@ -24,7 +24,6 @@ import org.scalatest.funsuite.AnyFunSuite // scalastyle:ignore funsuite
 import org.apache.spark.connect.proto
 import org.apache.spark.sql.Column
 import org.apache.spark.sql.connect.ColumnNodeToProtoConverter
-import org.apache.spark.util.TestEnvHelper
 
 /**
  * The basic testsuite the client tests should extend from.
@@ -32,8 +31,16 @@ import org.apache.spark.util.TestEnvHelper
 trait ConnectFunSuite
     extends AnyFunSuite // scalastyle:ignore funsuite
     with BeforeAndAfterAll
-    with BeforeAndAfterEach
-    with TestEnvHelper {
+    with BeforeAndAfterEach {
+
+  // Borrowed from SparkFunSuite
+  protected def getWorkspaceFilePath(first: String, more: String*): Path = {
+    if (!(sys.props.contains("spark.test.home") || sys.env.contains("SPARK_HOME"))) {
+      fail("spark.test.home or SPARK_HOME is not set.")
+    }
+    val sparkHome = sys.props.getOrElse("spark.test.home", sys.env("SPARK_HOME"))
+    java.nio.file.Paths.get(sparkHome, first +: more: _*)
+  }
 
   protected def baseResourcePath: Path = {
     getWorkspaceFilePath(
